requirements.txt
#------------------------------------------------------------------------------
#--- !!!   IMPORTANT   !!!
#    This is a Collection of all Project Dependencies.
#    !!! DO NOT REMOVE !!!
#------------------------------------------------------------------------------
#--- Upgrade all
#    pip freeze --local | grep -v "^\-e" | cut -d = -f 1 | xargs -n1 pip install -U

#------------------------------------------------------------------------------
# --- "#"
Django==4.1.5
PyPDF2==3.0.1
XlsxWriter==3.0.9

#------------------------------------------------------------------------------
# --- "A"
aioredis==2.0.1

# -----------------------------------------------------------------------------
# --- "B"
boto3==1.26.100

# -----------------------------------------------------------------------------
# --- "C"
channels==4.0.0
channels-redis==4.0.0
click==8.1.3
coverage==7.2.7

# -----------------------------------------------------------------------------
# --- "D"
djangorestframework==3.14.0
djangorestframework-jsonp==1.0.2
djangorestframework-jwt==1.11.0
djoser==2.1.0
dramatiq==1.14.2

# -----------------------------------------------------------------------------
# --- "django-*"
django-admin-rangefilter==0.10.0
django-filter==23.2
django-grappelli==3.0.6
django-redis==5.2.0
django-rest-framework-proxy==1.6.0
django-storages==1.13.2

#------------------------------------------------------------------------------
# --- "E"

#------------------------------------------------------------------------------
# --- "F"

#------------------------------------------------------------------------------
# --- "G"
gunicorn==20.1.0

#------------------------------------------------------------------------------
# --- "H"
httpx==0.24.0

#------------------------------------------------------------------------------
# --- "I"
ipython==8.13.2

#------------------------------------------------------------------------------
# --- "J"
jwt==1.3.1

#------------------------------------------------------------------------------
# --- "K"

#------------------------------------------------------------------------------
# --- "L"

#------------------------------------------------------------------------------
# --- "M"
markdown==3.4.3
mock==5.1.0

#------------------------------------------------------------------------------
# --- "N"
newrelic==8.7.0

#------------------------------------------------------------------------------
# --- "O"
openapi-spec-validator==0.5.6

#------------------------------------------------------------------------------
# --- "P"
paramiko==3.1.0
pdfkit==1.0.0
pendulum==2.1.2
pep8==1.7.1
pep257==0.7.0
phonenumbers==8.13.8
prance==0.22.11.4.0
psycopg2-binary==2.9.5
pycodestyle==2.5.0
pylint==2.15.10
pylint-django==2.5.3
pytest==7.4.0

# -----------------------------------------------------------------------------
# --- "python-*"
python-decouple==3.8

#------------------------------------------------------------------------------
# --- "Q"

#------------------------------------------------------------------------------
# --- "R"
rel==0.4.8
requests==2.28.2

# -----------------------------------------------------------------------------
# --- "S"

#------------------------------------------------------------------------------
# --- "T"
termcolor==2.3.0

#------------------------------------------------------------------------------
# --- "U"
uvicorn[standard]

#------------------------------------------------------------------------------
# --- "V"
vulture==2.7  # Find dead Code.

#------------------------------------------------------------------------------
# --- "W"
websocket-client==1.5.1

#------------------------------------------------------------------------------
# --- "X"
xlrd==2.0.1

#------------------------------------------------------------------------------
# --- "Y"

#------------------------------------------------------------------------------
# --- "Z"
